+++ Trionic5Tools/Trionic5Anomalies.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Text;
 using System.Data;
 using System.IO;
+using CommonSuite;
 
 namespace Trionic5Tools
 {
@@ -94,7 +95,7 @@ namespace Trionic5Tools
             {
                 if (sh.Varname == symbolname)
                 {
-                    retval = sh.Flash_start_address - m_fileInformation.Filelength;
+                    retval = (int)sh.Flash_start_address - m_fileInformation.Filelength;
                     while (retval > m_fileInformation.Filelength) retval -= m_fileInformation.Filelength;
                 }
             }
